+++ dsipts/data_structure/d1_layers/multi_source_csv.py
@@ -23,94 +23,6 @@ logging.basicConfig(
 logger = logging.getLogger(__name__)
 
 
-def extend_time_df(df, time_col, freq, group_cols=None, max_length=None):
-    """
-    Extend a dataframe to ensure regular time intervals.
-
-    Args:
-        df: Input dataframe containing time series data
-        time_col: Column name containing time information
-        freq: Frequency to use for extending the dataframe
-        group_cols: Optional list of columns identifying groups
-        max_length: Optional maximum length for the extended dataframe
-
-    Returns:
-        DataFrame extended with regular time intervals with all original columns preserved
-    """
-    if len(df) == 0:
-        return df
-
-    # Group by group columns if provided
-    if group_cols:
-        # Process each group separately
-        result_dfs = []
-        for group_key, group_df in df.groupby(group_cols):
-            extended_group = _extend_single_group(group_df, time_col, freq, max_length)
-            # Add group columns back
-            if isinstance(group_key, tuple):
-                for i, col in enumerate(group_cols):
-                    extended_group[col] = group_key[i]
-            else:
-                extended_group[group_cols[0]] = group_key
-            result_dfs.append(extended_group)
-        return pd.concat(result_dfs, ignore_index=True)
-    else:
-        return _extend_single_group(df, time_col, freq, max_length)
-
-
-def _extend_single_group(df, time_col, freq, max_length=None):
-    """Extend a single group's dataframe to ensure regular time intervals."""
-    if len(df) == 0:
-        return df
-
-    # Sort by time column
-    df = df.sort_values(time_col).reset_index(drop=True)
-
-    # Get min and max time
-    min_time = df[time_col].min()
-    max_time = df[time_col].max()
-
-    # Create complete time range
-    if pd.api.types.is_datetime64_dtype(df[time_col]):
-        # For datetime columns
-        time_range = pd.date_range(start=min_time, end=max_time, freq=freq)
-    else:
-        # For numeric columns
-        if isinstance(freq, (int, float)):
-            time_range = np.arange(min_time, max_time + freq, freq)
-        else:
-            # If freq is not numeric, try to infer step size
-            time_diffs = np.diff(df[time_col].dropna())
-            if len(time_diffs) > 0:
-                step = np.median(time_diffs)
-                time_range = np.arange(min_time, max_time + step, step)
-            else:
-                time_range = df[time_col].values
-
-    # Limit length if specified
-    if max_length and len(time_range) > max_length:
-        time_range = time_range[:max_length]
-
-    # Create result dataframe
-    result = pd.DataFrame({time_col: time_range})
-
-    # Add other columns and merge with original data
-    for col in df.columns:
-        if col != time_col:
-            result[col] = np.nan
-
-    # Merge with original data to fill in existing values
-    result = result.merge(df, on=time_col, how="left", suffixes=("", "_orig"))
-
-    # Fill in the merged values
-    for col in df.columns:
-        if col != time_col and f"{col}_orig" in result.columns:
-            result[col] = result[f"{col}_orig"]
-            result = result.drop(f"{col}_orig", axis=1)
-
-    return result
-
-
 class MultiSourceTSDataSet(BaseD1Layer):
     """
     Layer 1 (D1) dataset for multi-source time series data.
